# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Giant Monkey Robot Programming Challenge",
        "description": "PHP APIREST for the Giant Monkey Robot PHP Challenge",
        "contact": {
            "email": "patricio.quezada05@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/jobs/new": {
            "post": {
                "tags": [
                    "jobs"
                ],
                "summary": "New Job",
                "description": "ENDPOINT so that the client can create a new job.",
                "operationId": "newJob",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer <Token>"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Priority and Command of the Job",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "priority",
                                    "command"
                                ],
                                "properties": {
                                    "priority": {
                                        "type": "string",
                                        "format": "priority",
                                        "example": "low"
                                    },
                                    "command": {
                                        "type": "string",
                                        "format": "command",
                                        "example": "2+2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Job Dispatched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "format": "success",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "string",
                                            "format": "data",
                                            "example": "Job details"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dispatched job {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/jobs/get": {
            "get": {
                "tags": [
                    "jobs"
                ],
                "summary": "Get Job for procesing",
                "description": "ENDPOINT so that the Job Processor can get a Priority Accepted Job and execute his command",
                "operationId": "getJob",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer <Token>"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Job Processed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "format": "success",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "string",
                                            "format": "data",
                                            "example": "Job details"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Processed job {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "No job to process",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "format": "success",
                                            "example": "false"
                                        },
                                        "data": {
                                            "type": "string",
                                            "format": "data",
                                            "example": "No Data"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No job to process"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/jobs/{id}": {
            "get": {
                "tags": [
                    "jobs"
                ],
                "summary": "Get Status of a Job with an specific ID",
                "description": "ENDPOINT so that the Client can check the status of a Job",
                "operationId": "getJob",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer <Token>"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Job Processed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "format": "success",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "string",
                                            "format": "data",
                                            "example": "Job details"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Processed job {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "No job found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "format": "success",
                                            "example": "false"
                                        },
                                        "data": {
                                            "type": "string",
                                            "format": "data",
                                            "example": "No Data"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No job found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ],
...
}